+++ eveuniverse/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "eveuniverse.apps.EveuniverseConfig"
 
-__version__ = "0.6.0"
+__version__ = "0.6.1"
 __title__ = "Eve Universe"